refactor: centralize adapter api helpers
